🧠 refactor: dto 위치를 각 엔티티에 맞는 위치로 이동
issue: #40
